Оптимизaции

+++ Entities/User.py
@@ -23,7 +23,6 @@ def create():
     sql = """INSERT INTO User (username, about, name, email, isAnonymous) VALUES \
         (%(username)s, %(about)s, %(name)s, %(email)s, %(isAnonymous)s);"""
     args = {'username': username, 'about': about, 'name': name, 'email': email, 'isAnonymous': is_anonymous}
-    db = MyDatabase()
 
     try:
         db.execute(sql, args, True)
@@ -58,32 +57,30 @@ def details():
 
 @module.route("/follow/", methods=["POST"])
 def follow():
-    request_body = request.json
-
-    follower = request_body.get('follower')
-    followee = request_body.get('followee')
-
-    sql = """INSERT INTO Follower (follower, following) VALUES (%(follower)s, %(following)s);"""
-    args = {'follower': follower, 'following': followee}
-    db = MyDatabase()
-    db.execute(sql, args, True)
-    user_dict = get_user_dict(follower)
-    return json.dumps({"code": 0, "response": user_dict}, indent=4)
+    return follow_method(False)
 
 
 @module.route("/unfollow/", methods=["POST"])
 def unfollow():
+    return follow_method(True)
+
+
+def follow_method(do_unfollow):
     request_body = request.json
 
     follower = request_body.get('follower')
     followee = request_body.get('followee')
 
-    sql = """DELETE FROM Follower WHERE follower = %(follower)s AND following = %(following)s;"""
+    if not do_unfollow:
+        sql = """INSERT INTO Follower (follower, following) VALUES (%(follower)s, %(following)s);"""
+    else:
+        sql = """DELETE FROM Follower WHERE follower = %(follower)s AND following = %(following)s;"""
+
     args = {'follower': follower, 'following': followee}
-    db = MyDatabase()
+
     db.execute(sql, args, True)
-    user = get_user_dict(follower)
-    return json.dumps({"code": 0, "response": user}, indent=4)
+    user_dict = get_user_dict(follower)
+    return json.dumps({"code": 0, "response": user_dict}, indent=4)
 
 
 @module.route("/listPosts/", methods=["GET"])
@@ -106,16 +103,12 @@ def list_posts():
 def update_profile():
     request_body = request.json
 
-    about = request_body.get('about')
-    about = try_encode(about)
-    user = request_body.get('user')
-    user = try_encode(user)
-    name = request_body.get('name')
-    name = try_encode(name)
+    about = try_encode(request_body.get('about'))
+    user = try_encode(request_body.get('user'))
+    name = try_encode(request_body.get('name'))
 
     sql = """UPDATE User SET about = %(about)s, name = %(name)s WHERE email = %(email)s;"""
     args = {'about': about, 'name': name, 'email': user}
-    db = MyDatabase()
     db.execute(sql, args, True)
     user = get_user_dict(user)
     return json.dumps({"code": 0, "response": user}, indent=4)
@@ -170,7 +163,6 @@ def list_followers_method(is_following):
     sql += """ = %(email)s {since_value} {order_value} {limit_value};""".format(
         since_value=since_sql, order_value=order_sql, limit_value=limit_sql)
 
-    db = MyDatabase()
     user_list_sql = db.execute(sql, {'email': email})
     if not user_list_sql:
         return json.dumps({"code": 1, "response": "Empty set"}, indent=4)
